Borrado funciona (20 min)

+++ app/Http/Controllers/CategoryController.php
@@ -8,15 +8,19 @@ use Illuminate\Support\Facades\Validator;
 use DB;
 use Illuminate\Support\Facades\Auth;
 
-class CategoryController extends Controller { 
-/*     public function showAll($categorys) {
+class CategoryController extends Controller
+{
+    /*     public function showAll($categorys) {
         $categorys = Category::paginate(7);
         return view('admin.category', ['category' => $categorys]);
     } */
 
-    public function create(Request $request) {
+    public function create($name)
+    {
+        error_log('aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaabbbbbbbbbbbbbb');
+        /*
         $validator = Validator::make($request->all(), [
-            'categoria' => 'required',
+            'search-category' => 'required',
         ]);
         if ($validator->fails()) {
             return redirect(route('category'))
@@ -24,59 +28,48 @@ class CategoryController extends Controller {
                 ->withInput();
         }
         $inputs = $validator->validated();
-        $nombre = $request->input('categoria');
-        if(!$this->buscar($nombre, $request)) {
+        $nombre = $inputs['search-category'];
+        if (!$this->buscar($nombre)) {
             $new_category = new Category;
             //$new_category = Category::find($request->input('category_id'));
             $new_category->category = $nombre;
             $new_category->save();
         }
-        return redirect()->action([CategoryController::class, 'search'])->withInput();
-    }
-    
-    public function updateCategory(Request $request) {
-
+        */
+        return redirect()->action([CategoryController::class, 'showAll'])->withInput();
     }
 
-    public function buscar($nombre, Request $request) {
+    public function showAll()
+    {
+        $categorys = Category::all();
+        return view('admin.category', ['categorys' => $categorys]);
+    }
 
-        //$category =  Category::where('category', '==', $nombre)->withQueryString();
-        try{
-            // $category =  Category::where('categoria', "=", $nombre);
-            $category =  Category::where('categoria');
-/*             $categorys = Category::all();
-            foreach ($categorys as $item) {
-                dd($item->attributes);
-            } */
-            foreach ($request->query() as $item) {
-                if ($category == $item) {
-                    return true;
-                } else {
-                    return false;
-                }
-            }
-        } catch (QueryException $e) {
+    public function buscar($nombre)
+    {
+        $categorys = Category::where('category', '=', $nombre)->get();
+        $existe = $categorys->count();
+        if ($existe == 1) {
+            return true;
+        } else {
             return false;
         }
     }
 
-    public function search(Request $request) {
-        $nombre = $request->input('name');
-        if($nombre == null) {
-            $categorys = Category::paginate(7);
-        } else {
-            $categorys =  Category::where('name', 'LIKE', '%'.$nombre.'%')->orderBy('name', 'desc')->paginate(7)->withQueryString();
-        }
-        return view('admin.category', ['categorys' => $categorys]);
-    }
+    public function delete(Request $request)
+    {
+        $categorys = $request->all();
 
-    public function delete(Request $request) {
-        $categorys = json_decode($request->input('categorys'));
+        $i = 0;
         foreach ($categorys as $id) {
+            if ($i == 0) {
+                $i++;
+                continue;
+            }
             $category = Category::find($id);
             $category->delete();
         }
 
-        return back()->withInput();
+        return redirect()->action([CategoryController::class, 'showAll'])->withInput();
     }
 }
